+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog of SLEPc versions
 
-## [unreleased]
+## [3.21] - 2024-03-30
 
 ### Added
 
@@ -459,7 +459,8 @@
 - Classical and modified Gram-Schmidt orthogonalization use iterative refinement, with user options
   for parameter adjustment.
 
-[unreleased]: https://gitlab.com/slepc/slepc/compare/v3.20...main
+[unreleased]: https://gitlab.com/slepc/slepc/compare/v3.21...main
+[3.21]: https://gitlab.com/slepc/slepc/compare/v3.20...v3.21
 [3.20]: https://gitlab.com/slepc/slepc/compare/v3.19...v3.20
 [3.19]: https://gitlab.com/slepc/slepc/compare/v3.18...v3.19
 [3.18]: https://gitlab.com/slepc/slepc/compare/v3.17...v3.18